// inexus/types.ts
export type INexusConfig = {

    /**
     * api key
     */
    key: string

    /**
     * base url
     */
    base_url: string

    /**
     * model name
     */
    model?: string

    /**
     * system message
     */
    system_message?: string

    /**
     * content types
     */
    content_types: ('text' | 'image')[]

    /**
     * How long is the maximum loaded dialogue
     */
    max_time?: number

    /**
     * Maximum number of dialogues loaded
     */
    max_count?: number

    /**
     * max tokens
     */
    max_tokens?: number

    /**
     * max response tokens
     */
    max_response_tokens?: number

    /**
     * options
     */
    options?: {
        /**
         * org
         */
        org?: string

        /**
         * origin config
         */
        config?: any

        /**
         * cookie
         */
        cookie?: string
    } & { [key: string]: any }
}